+++ lib/phronomy/filter.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# Convenience require for Filter sub-classes.
+require_relative "filter/base"